requirements/production.txt
celery>=4.0.0
coreapi==2.2.4
django>=1.10
django-mail-queue==3.0.0
django-sendfile==0.3.11
enum34==1.1.6
html2text==2016.9.19
lxml==3.7.2
pathlib2==2.2.0
pygments==2.1.3
pyxb==1.2.5
redis==2.10.5
djangorestframework==3.5.4
